HAPP-1332: Create FetchLabOrdersUseCase
